backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via backend (pyproject.toml)
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   starlette
attrs==25.4.0
    # via aiohttp
bcrypt==5.0.0
    # via passlib
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via click
cryptography==46.0.3
    # via
    #   google-auth
    #   python-jose
distro==1.9.0
    # via google-genai
docstring-parser==0.17.0
    # via google-cloud-aiplatform
ecdsa==0.19.1
    # via python-jose
fastapi==0.128.0
    # via backend (pyproject.toml)
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.29.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.48.0
    # via
    #   backend (pyproject.toml)
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform==1.134.0
    # via backend (pyproject.toml)
google-cloud-bigquery==3.40.0
    # via google-cloud-aiplatform
google-cloud-core==2.5.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.16.0
    # via google-cloud-aiplatform
google-cloud-storage==3.8.0
    # via google-cloud-aiplatform
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.60.0
    # via google-cloud-aiplatform
google-resumable-media==2.8.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via google-cloud-resource-manager
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   langgraph-sdk
    #   langsmith
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.2.6
    # via backend (pyproject.toml)
langchain-core==1.2.7
    # via
    #   langchain
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langgraph==1.0.6
    # via
    #   backend (pyproject.toml)
    #   langchain
langgraph-checkpoint==4.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.6
    # via langgraph
langgraph-sdk==0.3.3
    # via langgraph
langsmith==0.6.4
    # via langchain-core
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
orjson==3.11.5
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   langchain-core
    #   langsmith
passlib==1.7.4
    # via backend (pyproject.toml)
pillow==12.1.0
    # via backend (pyproject.toml)
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.27.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==6.33.4
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-dotenv==1.2.1
    # via backend (pyproject.toml)
python-jose==3.5.0
    # via backend (pyproject.toml)
python-multipart==0.0.21
    # via backend (pyproject.toml)
pyyaml==6.0.3
    # via langchain-core
requests==2.32.5
    # via
    #   google-api-core
    #   google-auth
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via google-genai
starlette==0.50.0
    # via fastapi
tenacity==9.1.2
    # via
    #   google-genai
    #   langchain-core
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
    #   grpcio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
uuid-utils==0.14.0
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.40.0
    # via backend (pyproject.toml)
websockets==15.0.1
    # via google-genai
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
